requirements.txt
phidata
openai
python-pptx
